# api/openapi.yaml
openapi: "3.1.0"
info:
  title: AI Gateway API
  description: |
    AI Gateway 是一个统一的 API 网关，用于管理和代理 AI 模型提供商（如 OpenAI）的请求。
    它提供租户管理、API Key 认证、用量记录等功能。

    ## 快速开始

    1. 启动服务：`docker compose up -d`
    2. 创建租户：`POST /api/tenants`
    3. 创建 API Key：`POST /api/keys`
    4. 使用 Key 调用代理 API：`POST /v1/chat/completions`

    ## 认证方式

    代理 API（`/v1/*`）需要 Bearer Token 认证：
    ```
    Authorization: Bearer sk-xxxxxxxxxxxxxxxx
    ```

    管理 API（`/api/*`）当前为内部使用，后续将增加管理认证。
  version: "0.1.0"
  contact:
    name: AI Gateway Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
...
  /v1/models:
    get:
      operationId: listModels
      summary: 模型列表
      description: 获取可用的 AI 模型列表
      tags: [代理 API]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 模型列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ==========================================================================
  # 管理 API — 租户
  # ==========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-
      description: "格式: `Bearer sk-{prefix}{secret}`"
...
  schemas:
...
    ModelListResponse:
      type: object
      properties:
        object:
          type: string
          example: "list"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Model"
    Model:
      type: object
      properties:
        id:
          type: string
          example: "gpt-4"
        object:
          type: string
          example: "model"
        created:
          type: integer
        owned_by:
          type: string
          example: "openai"

    # -------------------------------------------------------------------------
    # 错误响应
    # -------------------------------------------------------------------------
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              description: 错误描述
              example: "invalid request body"
            type:
              type: string
              description: 错误类型
              example: "request_error"
            code:
              type: integer
              description: HTTP 状态码
              example: 400
          required: [message, type, code]
      required: [error]
  responses:
...
    Unauthorized:
      description: 未认证或认证失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...